book_service/requirements.txt
fastapi
uvicorn
sqlalchemy
confluent-kafka
pydantic
PyMySQL
kafka-python
cryptography